+++ chat/ConversationView.vue
@@ -1042,4 +1042,13 @@
             }
         }
     }
+
+    .user-avatar {
+      max-height: 1.2em;
+      min-height: 1.2em;
+      margin-right: 2px !important;
+      margin-top: 0;
+      min-width: 1.2em;
+      max-width: 1.2em;
+    }
 </style>